pyproject.toml
[project]
name = "au-sys-storage-service"
version = "0.1.0"
description = "Australis Systems Storage Service sub-module"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "beanie>=1.26.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "sqladmin>=0.16.0",
    "motor>=3.4.0",
    "fastapi>=0.100.0",
    "redis>=5.0.0",
    "redisvl>=0.3.0",
    "structlog>=24.1.0",
    "tinydb>=4.8.0",
    "aiosqlite>=0.19.0",
    "aiofiles>=23.2.1",
    "aioboto3>=12.0.0",
    "azure-storage-blob>=12.19.0",
    "google-cloud-storage>=2.13.0",
    "neo4j>=5.14.0",
    "falkordb>=1.0.0",
    "qdrant-client>=1.7.0",
    "httpx>=0.25.0",
    "mcp>=0.1.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0"
]

[project.scripts]
au-sys-storage-mcp = "au_sys_storage.stdio_server:main_sync"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["au_sys_storage"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
